Renaming bitstring_classes.py and moving some stuff to bitstore.py.
