+++ tests/test_bitstore.py
@@ -4,7 +4,7 @@ import unittest
 import sys
 sys.path.insert(0, '..')
 import bitstring
-from bitstring.bitstring_classes import BitStore, _offset_slice_indices_lsb0
+from bitstring.classes import BitStore, _offset_slice_indices_lsb0
 
 
 class BasicFunctionality(unittest.TestCase):
